Partial rewrite of scenario handling in SP menu
Not even pretending that splitting this into small chunks helps
understand any of the changes, simply because it's still horrible.

New bugs introduced due to changed scenario structure: Now require
metadata previously stored in root level to be defined in an explicit
`metadata` dictionary instead. This prevents parsing entire files
multiple times when querying difficulty, author, description and so on.

As for backwards compatibility, at least the map file should be detected
for old-style scenarios. No guarantees or promises about anything else.

Updates to the english parent scenarios (including the ones in tests/)
are contained in this commit. Updates to all other languages happen in a
different commit by running the new scenario translation script.